represent arrays with serialized xarray
