+++ paml/execution_engine.py
@@ -3,7 +3,7 @@ from typing import List
 import uuid
 import datetime
 import logging
-
+import pandas as pd
 import graphviz
 
 import paml
@@ -514,119 +514,3 @@ def protocol_execution_to_dot(self):
 
     return dot
 paml.ProtocolExecution.to_dot = protocol_execution_to_dot
-
-
-def call_behavior_execution_compute_output(self, parameter):
-    """
-    Get parameter value from call behavior execution
-    :param self:
-    :param parameter: output parameter to define value
-    :return: value
-    """
-    primitive = self.node.lookup().behavior.lookup()
-    call = self.call.lookup()
-    inputs = [x for x in call.parameter_values if x.parameter.lookup().property_value.direction == uml.PARAMETER_IN]
-    value = primitive.compute_output(inputs, parameter)
-    return value
-paml.CallBehaviorExecution.compute_output = call_behavior_execution_compute_output
-
-def primitive_compute_output(self, inputs, parameter):
-    """
-    Compute the value for parameter given the inputs. This default function will be overridden for specific primitives.
-    :param self:
-    :param inputs: list of paml.ParameterValue
-    :param parameter: Parameter needing value
-    :return: value
-    """
-
-    l.debug(f"Computing the output of primitive: {self.identity}, parameter: {parameter.name}")
-
-    if self.identity == 'https://bioprotocols.org/paml/primitives/sample_arrays/EmptyContainer' and \
-        parameter.name == "samples" and \
-        parameter.type == 'http://bioprotocols.org/paml#SampleArray':
-        # Make a SampleArray
-        for input in inputs:
-            i_parameter = input.parameter.lookup().property_value
-            value = input.value
-            if i_parameter.name == "specification":
-                spec = value.value.lookup().value if isinstance(value, uml.LiteralReference) else value.value
-        contents = self.initialize_contents()
-        name = f"{parameter.name}"
-        sample_array = paml.SampleArray(name=name,
-                                   container_type=spec,
-                                   contents=contents)
-        return sample_array
-    elif self.identity == 'https://bioprotocols.org/paml/primitives/sample_arrays/PlateCoordinates' and \
-        parameter.name == "samples" and \
-        parameter.type == 'http://bioprotocols.org/paml#SampleCollection':
-        for input in inputs:
-            i_parameter = input.parameter.lookup().property_value
-            value = input.value
-            if i_parameter.name == "source":
-                source = value.value.lookup() if isinstance(value, uml.LiteralReference) else value.value
-            elif i_parameter.name == "coordinates":
-                coordinates = value.value.lookup().value if isinstance(value, uml.LiteralReference) else value.value
-
-        mask = paml.SampleMask(source=source,
-                               mask=coordinates)
-        return mask
-    elif self.identity == 'https://bioprotocols.org/paml/primitives/spectrophotometry/MeasureAbsorbance' and \
-        parameter.name == "measurements" and \
-        parameter.type == 'http://bioprotocols.org/paml#SampleData':
-        for input in inputs:
-            i_parameter = input.parameter.lookup().property_value
-            value = input.value
-            if i_parameter.name == "samples":
-                samples = value.value.lookup() if isinstance(value, uml.LiteralReference) else value.value
-
-        sample_data = paml.SampleData(from_samples=samples)
-        return sample_data
-    else:
-        return f"{parameter.name}"
-paml.Primitive.compute_output = primitive_compute_output
-
-def empty_container_initialize_contents(self):
-    if self.identity == 'https://bioprotocols.org/paml/primitives/sample_arrays/EmptyContainer':
-        row_col_pairs = coordinate_rect_to_row_col_pairs("A1:H12")
-        aliquots = [uml.literal(f"{num2col(c+1)}{r}") for (r, c) in row_col_pairs]
-        elements = [uml.OrderedPropertyValue(index=i, property_value=aliquot)
-                    for i, aliquot in enumerate(aliquots)]
-        contents = paml.PrimitiveArray(elements=elements)
-    else:
-        raise Exception(f"Cannot initialize contents of: {self.identity}")
-    return contents
-paml.Primitive.initialize_contents = empty_container_initialize_contents
-
-def protocol_execution_set_data(self, data):
-    """
-    Overwrite execution trace values based upon values provided in data
-    """
-    execution_data = self.get_data()
-    for k, v in data.items():
-        execution_data[k] = v
-paml.ProtocolExecution.set_data = protocol_execution_set_data
-
-def protocol_execution_get_data(self):
-    """
-    Gather outputs from all CallBehaviorExecutions into a dict
-    """
-    def output_value(o):
-        return o.value.value.lookup() if isinstance(o.value, uml.LiteralReference) else o.value.value
-
-    calls = [e for e in self.executions if isinstance(e, paml.CallBehaviorExecution)]
-    data = {o.identity: output_value(o)
-            for e in calls
-            for o in e.get_outputs()
-            if isinstance(output_value(o), paml.SampleData)}
-
-    return data
-paml.ProtocolExecution.get_data = protocol_execution_get_data
-
-def activity_node_execution_get_outputs(self):
-    return []
-paml.ActivityNodeExecution.get_outputs = activity_node_execution_get_outputs
-
-def call_behavior_execution_get_outputs(self):
-    return [x for x in self.call.lookup().parameter_values
-              if x.parameter.lookup().property_value.direction == uml.PARAMETER_OUT]
-paml.CallBehaviorExecution.get_outputs = call_behavior_execution_get_outputs
